migre venta

+++ src-tauri/src/mods/config.rs
@@ -6,7 +6,7 @@ use std::sync::Arc;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Config {
-    politica_redondeo: f64,
+    politica_redondeo: f32,
     formato_producto: Formato,
     modo_mayus: Mayusculas,
     cantidad_productos: u8,
@@ -63,7 +63,7 @@ impl Config {
         }
     }
     pub fn build(
-        politica_redondeo: f64,
+        politica_redondeo: f32,
         formato_producto: &str,
         modo_mayus: &str,
         cantidad_productos: u8,
@@ -94,7 +94,7 @@ impl Config {
     pub fn medios_pago(&self) -> &Vec<Arc<str>> {
         &self.medios_pago
     }
-    pub fn politica(&self) -> f64 {
+    pub fn politica(&self) -> f32 {
         self.politica_redondeo
     }
     pub fn formato(&self) -> &Formato {
