+++ src-tauri/src/mods/pago.rs
@@ -29,12 +29,12 @@ impl MedioPago {
 pub struct Pago {
     int_id: i64,
     medio_pago: MedioPago,
-    monto: f64,
-    pagado: f64,
+    monto: f32,
+    pagado: f32,
 }
 
 impl Pago {
-    pub fn new(medio_pago: MedioPago, monto: f64, pagado: Option<f64>) -> Pago {
+    pub fn new(medio_pago: MedioPago, monto: f32, pagado: Option<f32>) -> Pago {
         let int_id = random();
         Pago {
             medio_pago,
@@ -46,7 +46,7 @@ impl Pago {
             },
         }
     }
-    pub fn build(int_id: i64, medio_pago: MedioPago, monto: f64, pagado: f64) -> Pago {
+    pub fn build(int_id: i64, medio_pago: MedioPago, monto: f32, pagado: f32) -> Pago {
         Pago {
             int_id,
             medio_pago,
@@ -60,13 +60,13 @@ impl Pago {
     pub fn medio(&self) -> Arc<str> {
         Arc::clone(&self.medio_pago.medio)
     }
-    pub fn monto(&self) -> f64 {
+    pub fn monto(&self) -> f32 {
         self.monto
     }
     pub fn id(&self) -> i64 {
         self.int_id
     }
-    pub fn pagado(&self) -> &f64 {
+    pub fn pagado(&self) -> &f32 {
         &self.pagado
     }
     pub fn def(db: &Pool<Sqlite>) -> Self {
